requirements.txt
# Core
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Predict.fun SDK
predict-sdk==0.0.12

# Web3 & Blockchain
web3==6.14.0
eth-account==0.10.0

# WebSocket
websocket-client==1.7.0
python-socketio==5.10.0

# Async
aiohttp==3.9.1
asyncio==3.4.3

# Utilities
pytz==2024.1
python-dateutil==2.8.2

# Logging
colorlog==6.8.0

# Telegram Bot
python-telegram-bot==20.7
